requirements.txt
aiofiles==0.7.0
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
decorator==5.0.9
fastapi==0.68.1
future==0.18.2
geographiclib==1.52
h11==0.12.0
httptools==0.2.0
idna==3.2
Jinja2==3.0.1
MarkupSafe==2.0.1
pydantic==1.8.2
python-dotenv==0.19.0
PyYAML==5.4.1
ratelim==0.1.6
requests==2.26.0
six==1.16.0
sodapy==2.1.0
starlette==0.14.2
typing-extensions==3.10.0.0
urllib3==1.26.6
uvicorn==0.15.0
uvloop==0.16.0
watchgod==0.7
websockets==9.1
